+++ backend/src/main/java/com/tesis/demo/dto/RolDto.java
@@ -1,6 +1,7 @@
 package com.tesis.demo.dto;
 
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class RolDto {
-    private Long id;
     @NotBlank
-    private String nombreRol;
+    @NotNull
+    private String nombre;
 }